Add hooks for setting bounds and permissions on exported pointers.
On a CHERI system, the pointers into the Microvium heap give full
read-write access to the entire heap.  This means that bounds errors (or
passing string pointers to functions that try to mutate their arguments)
will corrupt VM state.

The two hooks are:

 - Apply bounds to this pointer.
 - Remove the ability to load pointers or store through the pointer.

Both of these default to doing nothing, but no CHERI platforms they can
be a csetbounds / candperms instruction.